[Refactor] source codes
