+++ src/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_K = [2.6, 2.8, 3.0, 3.2]
