+++ src/sql/scoring_query.sql
